move atomic operations and atomic handler to separate modules
